Add job monitoring functionality and enhance job data model
- Introduced a new monitoring section to track job progress and status.
- Added `pid` and `last_updated` fields to the JobDataModel for better process management.
- Implemented functions to calculate progress based on job phases and log parsing.
- Enhanced database utility functions to support new fields in job data.
- Added logic for non-blocking PID checks to ensure job status accuracy.
- Improved log handling and display for better user experience in the monitoring section.
