+++ streamlit_app/models/job_data_model.py
@@ -18,6 +18,7 @@ class JobDataModel(BaseModel):
     start_time: float = Field(default_factory=time.time)
     end_time: Optional[float] = None
     process: Optional[Process] = None
+    pid: Optional[int] = None  # Store process ID for PID checks after Streamlit reruns
     status_queue: Optional[Any] = None  # Using Any to support both Queue types
     config: Dict[str, Any] = Field(default_factory=dict)
     output_final_file_path: Optional[str] = None
@@ -27,9 +28,19 @@ class JobDataModel(BaseModel):
     file_info: Dict[str, Any] = Field(default_factory=dict)
     log_messages: List[str] = Field(default_factory=list)
     max_progress: float = 0.0
+    last_updated: Optional[float] = Field(default_factory=time.time)
 
     class Config:
         """Pydantic model configuration."""
 
         arbitrary_types_allowed = True  # Needed for multiprocessing.Process and Queue
         validate_assignment = False
+
+    def model_post_init(self, __context: Any) -> None:
+        # Ensure last_updated is set on creation if not provided
+        if self.last_updated is None:
+            self.last_updated = time.time()
+
+    def touch(self) -> None:
+        """Updates the last_updated timestamp."""
+        self.last_updated = time.time()
